pyproject.toml
[tool.poetry]
name = "mediasoft-test"
version = "0.1.0"
description = "Test"
authors = ["u-shev <sheveleva.jul@gmail.com>"]
readme = "README.md"
packages = [{include = "mediasoft_test"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
django = "^4.2.7"
djangorestframework = "^3.14.0"
django-filter = "^23.3"
python-dotenv = "^1.0.0"
flake8 = "^6.1.0"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
